requirements.txt
appnope==0.1.0
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.6.0
bleach==3.1.0
boto3==1.9.155
botocore==1.12.196
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.2
django-letsencrypt==3.0.1
django-modelcluster==4.3
django-sendgrid-v5==0.8.0
django-sslserver==0.20
django-storages==1.7.1
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.9.2
djangorestframework-csv==2.1.0
docutils==0.14
draftjs-exporter==2.1.5
entrypoints==0.3
et-xmlfile==1.0.1
future==0.16.0
gunicorn==19.7.1
html5lib==1.0.1
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.3
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.5.0
nbformat==4.4.0
nose==1.3.7
notebook==5.7.8
numpy==1.16.4
odfpy==1.4.0
openpyxl==2.6.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psycopg2==2.7.7
psycopg2-binary==2.8.2
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.6.1
python-http-client==3.1.0
pytz==2018.9
PyYAML==5.1.0
pyzmq==18.0.1
qtconsole==4.5.1
requests==2.21.0
s3transfer==0.2.0
scikit-learn==0.21.2
scipy==1.3.0
Send2Trash==1.5.0
sendgrid==6.0.5
six==1.12.0
sympy==1.4
tablib==0.13.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
unicodecsv==0.14.1
Unidecode==1.0.23
urllib3==1.24.2
wagtail==2.4
wagtail-robots==0.1.2
wagtail-schema.org==2.0.0
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==3.3.1
widgetsnbextension==3.4.2
Willow==1.1
xlrd==1.2.0
xlwt==1.3.0
